Analysis properly handles holes close to the edge
Previously, the strategy was simply to select a smaller box than we
could have gotten away with.  This doesn't actually fix the problem of
points near the edge, since it just changes what the edge is.

Now, anything within "padding" of the edge will be ignored (set to NaN).
This happens after the fit is calculated.